Reorganize survey agent/trigger functions on backend
