+++ functions/src/agent_participant.utils.ts
@@ -13,7 +13,7 @@ import {updateParticipantNextStage} from './participant.utils';
 import {initiateChatDiscussion} from './stages/chat.utils';
 import {completeProfile} from './stages/profile.utils';
 import {getAgentParticipantRankingStageResponse} from './stages/ranking.utils';
-import {getAgentParticipantSurveyStageResponse} from './stages/survey.utils';
+import {getAgentParticipantSurveyStageResponse} from './stages/survey.agent';
 import {
   getExperimenterData,
   getFirestoreParticipantRef,
